Move more code out of global utils
