+++ apertium_apy/utils/__init__.py
@@ -6,7 +6,6 @@ from subprocess import Popen, PIPE
 from threading import Thread
 
 from apertium_apy.missingdb import timedelta_to_milliseconds
-from apertium_apy.utils.wiki import wiki_get_page, wiki_edit_page, wiki_add_text
 
 iso639Codes = {'abk': 'ab', 'aar': 'aa', 'afr': 'af', 'aka': 'ak', 'sqi': 'sq', 'amh': 'am', 'ara': 'ar', 'arg': 'an', 'hye': 'hy', 'asm': 'as', 'ava': 'av', 'ave': 'ae', 'aym': 'ay', 'aze': 'az', 'bam': 'bm', 'bak': 'ba', 'eus': 'eu', 'bel': 'be', 'ben': 'bn', 'bih': 'bh', 'bis': 'bi', 'bos': 'bs', 'bre': 'br', 'bul': 'bg', 'mya': 'my', 'cat': 'ca', 'cha': 'ch', 'che': 'ce', 'nya': 'ny', 'zho': 'zh', 'chv': 'cv', 'cor': 'kw', 'cos': 'co', 'cre': 'cr', 'hrv': 'hr', 'ces': 'cs', 'dan': 'da', 'div': 'dv', 'nld': 'nl', 'dzo': 'dz', 'eng': 'en', 'epo': 'eo', 'est': 'et', 'ewe': 'ee', 'fao': 'fo', 'fij': 'fj', 'fin': 'fi', 'fra': 'fr', 'ful': 'ff', 'glg': 'gl', 'kat': 'ka', 'deu': 'de', 'ell': 'el', 'grn': 'gn', 'guj': 'gu', 'hat': 'ht', 'hau': 'ha', 'heb': 'he', 'her': 'hz', 'hin': 'hi', 'hmo': 'ho', 'hun': 'hu', 'ina': 'ia', 'ind': 'id', 'ile': 'ie', 'gle': 'ga', 'ibo': 'ig', 'ipk': 'ik', 'ido': 'io', 'isl': 'is', 'ita': 'it', 'iku': 'iu', 'jpn': 'ja', 'jav': 'jv', 'kal': 'kl', 'kan': 'kn', 'kau': 'kr', 'kas': 'ks', 'kaz': 'kk', 'khm': 'km', 'kik': 'ki', 'kin': 'rw', 'kir': 'ky', 'kom': 'kv', 'kon': 'kg', 'kor': 'ko', 'kur': 'ku', 'kua': 'kj', 'lat': 'la', 'ltz': 'lb', 'lug': 'lg', 'lim': 'li', 'lin': 'ln', 'lao': 'lo', 'lit': 'lt', 'lub': 'lu', 'lav': 'lv', 'glv': 'gv', 'mkd': 'mk', 'mlg': 'mg', 'msa': 'ms', 'mal': 'ml', 'mlt': 'mt', 'mri': 'mi', 'mar': 'mr', 'mah': 'mh', 'mon': 'mn', 'nau': 'na', 'nav': 'nv', 'nob': 'nb', 'nde': 'nd', 'nep': 'ne', 'ndo': 'ng', 'nno': 'nn', 'nor': 'no', 'iii': 'ii', 'nbl': 'nr', 'oci': 'oc', 'oji': 'oj', 'chu': 'cu', 'orm': 'om', 'ori': 'or', 'oss': 'os', 'pan': 'pa', 'pli': 'pi', 'fas': 'fa', 'pol': 'pl', 'pus': 'ps', 'por': 'pt', 'que': 'qu', 'roh': 'rm', 'run': 'rn', 'ron': 'ro', 'rus': 'ru', 'san': 'sa', 'srd': 'sc', 'snd': 'sd', 'sme': 'se', 'smo': 'sm', 'sag': 'sg', 'srp': 'sr', 'gla': 'gd', 'sna': 'sn', 'sin': 'si', 'slk': 'sk', 'slv': 'sl', 'som': 'so', 'sot': 'st', 'azb': 'az', 'spa': 'es', 'sun': 'su', 'swa': 'sw', 'ssw': 'ss', 'swe': 'sv', 'tam': 'ta', 'tel': 'te', 'tgk': 'tg', 'tha': 'th', 'tir': 'ti', 'bod': 'bo', 'tuk': 'tk', 'tgl': 'tl', 'tsn': 'tn', 'ton': 'to', 'tur': 'tr', 'tso': 'ts', 'tat': 'tt', 'twi': 'tw', 'tah': 'ty', 'uig': 'ug', 'ukr': 'uk', 'urd': 'ur', 'uzb': 'uz', 'ven': 've', 'vie': 'vi', 'vol': 'vo', 'wln': 'wa', 'cym': 'cy', 'wol': 'wo', 'fry': 'fy', 'xho': 'xh', 'yid': 'yi', 'yor': 'yo', 'zha': 'za', 'zul': 'zu', 'hbs': 'sh', 'arg': 'an', 'pes': 'fa'}  # noqa: E501
 """
@@ -44,6 +43,15 @@ def to_alpha3_code(code):
         return iso639_codes_inverse[code] if code in iso639_codes_inverse else code
 
 
+def remove_dot_from_deformat(query, analyses):
+    """When using the txt format, a dot is added at EOF (also, double line
+    breaks) if the last part of the query isn't itself a dot"""
+    if not query[-1] == '.':
+        return analyses[:-1]
+    else:
+        return analyses
+
+
 def apertium(input, mode_dir, mode, formatting='txt'):
     p1 = Popen(['echo', input], stdout=PIPE)
     logging.getLogger().info('util.apertium({}, {}, {}, {})'
@@ -54,30 +62,6 @@ def apertium(input, mode_dir, mode, formatting='txt'):
     p1.stdout.close()
     output = p2.communicate()[0].decode('utf-8')
     return output
-
-
-def bilingual_translate(to_translate, mode_dir, mode):
-    p1 = Popen(['echo', to_translate], stdout=PIPE)
-    p2 = Popen(['lt-proc', '-b', mode], stdin=p1.stdout, stdout=PIPE, cwd=mode_dir)
-    p1.stdout.close()
-    output = p2.communicate()[0].decode('utf-8')
-    return output
-
-
-def remove_dot_from_deformat(query, analyses):
-    """When using the txt format, a dot is added at EOF (also, double line
-    breaks) if the last part of the query isn't itself a dot"""
-    if not query[-1] == '.':
-        return analyses[:-1]
-    else:
-        return analyses
-
-
-def strip_tags(analysis):
-    if '<' in analysis:
-        return analysis[:analysis.index('<')]
-    else:
-        return analysis
 
 
 def get_coverages(text, modes, penalize=False):
@@ -115,20 +99,3 @@ def scale_mt_log(status, time, t_info, key, length):
                                         length,
                                         'null',
                                         )
-
-
-def add_suggestion(s, suggest_url, edit_token, data):
-    content = wiki_get_page(s, suggest_url)
-    content = wiki_add_text(content, data)
-    edit_result = wiki_edit_page(s, suggest_url, content, edit_token)
-
-    try:
-        if edit_result['edit']['result'] == 'Success':
-            logging.info('Update of page %s' % (suggest_url))
-            return True
-        else:
-            logging.error('Update of page %s failed: %s' % (suggest_url,
-                                                            edit_result))
-            return False
-    except KeyError:
-        return False
